Extract handlers.py files
